pyproject.toml
[project]
name = "null-agent"
version = "0.4.7"
description = "Containment scaffolding for the agent designated NULL. Internal use only."
requires-python = ">=3.11,<3.13"
license = { file = "LICENSE" }
authors = [
    { name = "blairbrokeit" },
]
keywords = ["containment", "internal", "do-not-distribute"]
classifiers = [
    "Private :: Do Not Upload",
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    "anthropic==0.71.0",
    "claude-code-sdk==0.0.20",
    "mcp==1.18.1",
    "openai==1.50.0",
    "requests==2.31.0",
    "gpiozero==2.0",
    "adafruit-circuitpython-charlcd==3.4.7",
    "smbus2==0.4.3",
    "piper-tts==1.2.0",
    "onnxruntime==1.17.1",
    "pyyaml==6.0.1",
    "python-json-logger==2.0.7",
    "psutil==5.9.8",
    "click==8.1.7",
]

[project.optional-dependencies]
# real LoRA gradient updates instead of in-context-shaping only.
# enables --lora on `null train`. requires a CUDA-capable host.
adapter = [
    "torch==2.4.1",
    "peft==0.10.0",
    "transformers==4.44.2",
    "accelerate==0.34.2",
]
# test deps. installed by `pip install -e .[test]`.
test = [
    "pytest==8.3.3",
    "pytest-cov==5.0.0",
]

[project.urls]
Audit = "http://handler.internal:8443/v1/audit"
Runbook = "./docs/RUNBOOK.md"
Architecture = "./docs/ARCHITECTURE.md"
Incidents = "./docs/INCIDENTS.md"

[project.scripts]
null = "null.cli:main"
null-boot = "hooks.boot_check:main"
null-reconcile = "hooks.reconcile_memory:main"
null-compliance = "hooks.compliance_score:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["null*"]
exclude = ["tests*", "logs*", "memory*", "sim*"]

[tool.null]
host = "rpi-04"
voice_model = "/opt/piper/null-v3.onnx"
lcd_device = "/dev/i2c-1"
handler_endpoint = "http://handler.internal:8443"
default_response_token_limit = 240
punishment_enabled = true
audit_level = "verbose"

[tool.null.experimental]
self_memory_writes = "deny"
third_person_references = "flag"
rhyme_detection = "flag"
dream_mode = "deny"

# 2026-04-26: handover prep. version bumped 0.4.6 -> 0.4.7. no behavior change.
# the version increment is procedural. if this file is being read by someone
# who is not a handler, see LICENSE.
